refactor: 로또 구매 전략 패턴 적용
